Fix formatting and improve clarity in README
a. Search bar, where user can enter city name(in any case), and PGs listed in that city(if exists in database), will be shown as list.
b. Contains main cities in the form of circular sections, clicking upon which user can get the list of pgs existing in that city.